Add support for Visual C++ 2012, include debug info in release build.
The file AgtCtl_i.c is no longer delivered with Visual C++ 2012.
This file is needed to compile extensions\platform\windows\ole.
Workaround : copy this file from an older version of VC++, paste it in the ole directory.

Debug infos in release build : I keep all the optimizations, my goal is just to have a call stack with symbolic names, in case of crash.